backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# 认证相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# 环境变量管理
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP请求
httpx==0.25.2

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1